package ui renamed to package view
