+++ src/Main.java
@@ -1,4 +1,4 @@
-import ui.UserPage;
+import view.UserPage;
 
 public class Main {
     public static void main(String[] args) {
